requirements.txt
pandas>=2.1
numpy>=1.24
scikit-learn>=1.3
joblib>=1.3
python-dotenv>=1.0
Flask>=3.0
SQLAlchemy>=2.0
requests>=2.31
beautifulsoup4>=4.12
pytest>=8.0
transformers>=4.35
datasets>=2.14
torch>=2.0
google-api-python-client>=2.0
google-auth-oauthlib>=1.2
google-auth-httplib2>=0.2

